// src/Brekeke/types.ts
export interface Pbx {
  login(onres: () => void, onerr: (err: Error) => void): void
  close(): void

  debugLevel: number

  onClose?(): void
  onError?(err: Error): void
  notify_serverstatus?(e: PbxEvent['serverStatus']): void
  notify_status?(e: PbxEvent['userStatus']): void
  notify_park?(e: PbxEvent['park']): void
  notify_voicemail?(e: PbxEvent['voicemail']): void

  pal<K extends keyof PbxPal>(
    k: K,
    p: Parameters<PbxPal[K]>[0],
  ): Promise<Parameters<Parameters<PbxPal[K]>[1]>[0]>
}

export interface PbxEvent {
  serverStatus: {
    status: 'inactive'
  }
  userStatus: {
    status: string
    user: string
    talker_id: string
  }
  park: {
    park: string
    status: 'on' | 'off'
  }
  voicemail: {
    new: number
  }
}

export interface PbxPal {
  getProductInfo(
    p: null,
    onres: (i: { 'sip.wss.port': string }) => void,
    onerr: (err: Error) => void,
  ): void

  createAuthHeader(
    p: { username: string },
    onres: (authHeader: string) => void,
    onerr: (err: Error) => void,
  ): void

  getExtensions(
    p: {
      tenant: string
      pattern: '..*'
      type: 'user'
      limit: number
    },
    onres: (extensions: string[]) => void,
    onerr: (err: Error) => void,
  ): void

  getExtensionProperties<T extends string | string[]>(
    p: {
      tenant: string
      extension: T
      property_names: string[]
    },
    onres: (properties: T[]) => void,
    onerr: (err: Error) => void,
  ): void

  setExtensionProperties(
    p: {
      tenant: string
      extension: string
      properties: {
        p1_ptype?: string
        p2_ptype?: string
        p3_ptype?: string
        p4_ptype?: string
        pnumber: string
      }
    },
    onres: () => void,
    onerr: (err: Error) => void,
  ): void

  pnmanage(
    p: {
      command: string
      service_id: string
      application_id: string
      user_agent: string
      username: string
      device_id?: string
      endpoint?: string
      auth_secret?: string
      key?: string
    },
    onres: () => void,
    onerr: (err: Error) => void,
  ): void
}